<Frontend> modified the pdfReader component to the default react-pdf way
