+++ components/PdfReader.tsx
@@ -1,0 +1,32 @@
+"use client";
+import { useState } from "react";
+import { pdfjs, Document, Page } from "react-pdf";
+import "react-pdf/dist/esm/Page/AnnotationLayer.css";
+import "react-pdf/dist/esm/Page/TextLayer.css";
+
+pdfjs.GlobalWorkerOptions.workerSrc = `//unpkg.com/pdfjs-dist@${pdfjs.version}/legacy/build/pdf.worker.min.mjs`;
+
+export default function PdfReader() {
+  const pdfFile = "/mdlb.pdf";
+  const [numPages, setNumPages] = useState<number>();
+  const [pageNumber, setPageNumber] = useState<number>(1);
+
+  function onDocumentLoadSuccess({ numPages }: { numPages: number }): void {
+    setNumPages(numPages);
+  }
+
+  return (
+    <div>
+      <Document file={pdfFile} onLoadSuccess={onDocumentLoadSuccess}>
+        <Page pageNumber={pageNumber} />
+      </Document>
+      <p>
+        Page {pageNumber} of {numPages}
+      </p>
+      <div className="flex justify-around">
+        <button onClick={() => setPageNumber(pageNumber - 1)}>Back page</button>
+        <button onClick={() => setPageNumber(pageNumber + 1)}>Next Page</button>
+      </div>
+    </div>
+  );
+}
